IsOptionRequested method of DHCPv6Message
