+++ dhcpv4/dhcpv4.go
@@ -682,9 +682,8 @@ func (d *DHCPv4) ValidateOptions() {
 	}
 }
 
-// IsOptionRequested function takes a DHCPv4 message and an OptionCode, and
-// returns true if that option is within the requested options of the DHCPv4
-// message.
+// IsOptionRequested returns true if that option is within the requested
+// options of the DHCPv4 message.
 func (d *DHCPv4) IsOptionRequested(requested OptionCode) bool {
 	for _, optprl := range d.GetOption(OptionParameterRequestList) {
 		for _, o := range optprl.(*OptParameterRequestList).RequestedOpts {
